+++ src/main/java/blackjack/domain/result/ScoreComparator.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 public interface ScoreComparator {
