+++ src/test/java/blackjack/domain/DeckTest.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.card.CardGenerator;
+import blackjack.domain.card.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
